chatbot_project/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
contourpy==1.3.0
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.0.6
django-crispy-forms==2.3
django-heroku==0.3.1
django-modeladmin-reorder==0.3.1
django-nested-inline==0.4.6
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.14.0
flake8==7.1.0
fonttools==4.54.1
frozenlist==1.4.1
geopandas==1.0.1
google-api-core==1.34.1
google-auth==2.36.0
google-cloud-dialogflow==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.48.2
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
mysql-connector-python==9.1.0
mysqlclient==2.2.4
nltk==3.9.1
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.1
preshed==3.0.9
propcache==0.2.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==3.20.3
psycopg2==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pyogrio==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.2.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
queuelib==1.7.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rsa==4.9
Scrapy==2.11.2
seaborn==0.13.2
selenium==4.26.1
service-identity==24.1.0
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
sublime==0.1.0
thinc==8.3.2
tldextract==5.1.2
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
twilio==9.3.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
w3lib==2.2.1
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
whitenoise==6.8.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.15.4
zope.interface==6.4.post2
